Replaced ZET CPU with ao486 CPU
The ZET CPU had too many problems that needed to be fixed in order to
get it to work with Verilator. The ao486 is a 32 bit CPU and not much
work will be needed to get it to work with Verilator. Lets try using
ao486 to see if we can get it to work in our design.